+++ aslam_offline_calibration/kalibr/python/kalibr_camera_calibration/__init__.py
@@ -4,3 +4,4 @@ from .CameraCalibrator import *
 from .ObsDb import *
 from .MulticamGraph import *
 from .CameraIntializers import *
+from .CameraUtils import *
